add current photo
