+++ index.html
@@ -234,15 +234,6 @@
   </template>
 </body>
 <script src="js/main.js" type = "module"></script>
-<script src="js/util.js" type = "module"></script>
-<script src="js/data.js" type = "module"></script>
-<script src="js/render.js" type="module"></script>
-<script src="js/render-big-pictures.js" type="module"></script>
-<script src="js/form.js" type="module"></script>
-<script src="js/validate-form.js" type="module"></script>
 <script src="pristine/pristine.min.js"></script>
-<script src="js/scale.js" type="module"></script>
-<script src="js/filter.js" type="module"></script>
-<script src="js/fetch.js" type="module"></script>
 <script src="nouislider/nouislider.js"></script>
 </html>
